+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 **Bug Fixes**
 - Create block pool only in the child process.
 
+**Features**
+- To protect against accidental overwrites on data stored by block-cache on temp path, md5 sums will be validated on read. This feature can be enabled by using `--block-cache-strong-consistency` cli flag.
+
 ## 2.4.0 (Unreleased)
 **Features**
 - Added 'gen-config' command to auto generate the recommended blobfuse2 config file based on computing resources and memory available on the node. Command details can be found with `blobfuse2 gen-config --help`.